Emailer (#58)
* Started work on email out ** Had to upgrade sdk version to 19

Currently click on email in export, it opens up document picker to pick a document to send
Then sends.

** Email requires security setting change to allow unverified applications to access **

* Added specific email tile to send email with attachment choice.

* Added helper functions in Object Handler to get raw file content for LogInfo/Units/Tests for a given currentState.currentDocument/currentState.unitsList/currentState.testsLists

Created temp UI in export to send the contents of a document to Gmail to be emailed out.
Option to choose JSON or CSV format.
After sending email will take you back to application.

* Reworked Send email to be able to open gmail with chosen attachment.

Choose the document name and either csv or pdf format. Have dummy dropdown for both in a zip file which will be fixed later.

Doesn't work if the .csv or .pdf is not found. Will update when we put in project wrapper to have the ability to auto create pdf/csv files on demand, put into zip and attach if desired.

* Cleaned up code/commented out iterations of previous attempts

* Removed some old code
Changed pdf text name to not include date for the time being(Will version later)

* Minor code cleanup

* - caught unhandled exception when trying to send file that is not found
 - removed email tile
 - removed dummy option

* commented out trace statement